requirements.txt
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
PyYAML>=5.4.0
tqdm>=4.60.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.1.0
scipy>=1.5.0

# YOLOv5 dependencies
gitpython>=3.1.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0
